% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sig_score3.R
\name{calculate_sig_score3}
\alias{calculate_sig_score3}
\title{Title}
\usage{
calculate_sig_score3(
  eset,
  signature_zscore = feas_sig_op_zscore,
  signature_pca = feas_sig_op_pca,
  signature_ssgsea = feas_sig_op_ssgsea
)
}
\arguments{
\item{signature_ssgsea}{}
}
\description{
Title
}
