% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lira_data.R
\name{detect_miss}
\alias{detect_miss}
\title{Title Detect missing genes of signature gene list}
\usage{
detect_miss(eset, signature)
}
\arguments{
\item{eset}{expression set}

\item{signature}{a gene list}
}
\description{
Title Detect missing genes of signature gene list
}
