% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lira_data.R
\name{get_lira_data}
\alias{get_lira_data}
\title{Title}
\usage{
get_lira_data(
  id = "LIRA0001",
  data_type = c("tpm", "scale_tpm", "count"),
  path = "E:/03-NSCLC/19-NSCLC-LIRA/4-analysis/0-data/"
)
}
\arguments{
\item{id}{LIRA0001}

\item{data_type}{"tpm", "scale_tpm", "count"}

\item{path}{"E:/03-NSCLC/19-NSCLC-LIRA/4-analysis/0-data/"}
}
\description{
Title
}
