% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_heatmap.R
\name{lira_heatmap}
\alias{lira_heatmap}
\title{Title}
\usage{
lira_heatmap(
  eset,
  pdata = NULL,
  id_pdata = "ID",
  BOR = "BOR",
  palette_heatmap = 6,
  path = NULL
)
}
\arguments{
\item{BOR}{}

\item{palette_heatmap}{}

\item{path}{}
}
\description{
Title
}
