% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_model.R
\name{lira_model}
\alias{lira_model}
\title{Title}
\usage{
lira_model(
  eset,
  pdata = NULL,
  id_pdata = "ID",
  scale = FALSE,
  plot = FALSE,
  check_eset = FALSE
)
}
\arguments{
\item{check_eset}{}
}
\description{
Title
}
