% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_rnaseq_pipline.R
\name{lira_rnaseq_pipline}
\alias{lira_rnaseq_pipline}
\title{Title}
\usage{
lira_rnaseq_pipline(
  pat_id,
  eset = NULL,
  data_type = "tpm",
  palette = "nrc",
  path1 = NULL,
  path2 = NULL
)
}
\arguments{
\item{pat_id}{}

\item{eset}{}

\item{data_type}{}

\item{palette}{}

\item{path1}{}

\item{path2}{}
}
\description{
Title
}
\examples{
需要设置新数据的关键变量，否则后续画图会报错
寻找最佳的cutoff值
}
