% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_score_location.R
\name{lira_score_location}
\alias{lira_score_location}
\title{Location of lira_score of each patient}
\usage{
lira_score_location(
  score,
  ref_score = ref_lira_score,
  palette = "nrc",
  cols = NULL,
  palette_line = "jama",
  showplot = TRUE,
  path = NULL,
  panel = "OS"
)
}
\arguments{
\item{score}{score}

\item{ref_score}{}

\item{palette}{palette of response}

\item{cols}{}

\item{palette_line}{palette of line}

\item{showplot}{default is FALSE}

\item{path}{path to save result}

\item{panel}{}
}
\description{
Location of lira_score of each patient
}
