\name{PoissonEImages}
\docType{data}
\alias{PoissonEImages}
\title{LIRA Test Images}
\description{
  This data set provides three test images for the lira package. \code{PoissonE.obs} is a simulated observed image, \code{PoissonE.null} is a null model, and \code{PoissonE.flat} is a uniform background.
}
\usage{data(PoissonEImages)}
\format{32-by-32 matrices.}
\keyword{datasets}