% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_model.R
\name{lira_model}
\alias{lira_model}
\title{This function uses the LIRA model to predict new data and return the results.}
\usage{
lira_model(
  eset,
  pdata = NULL,
  id_pdata = "ID",
  scale = FALSE,
  check_eset = FALSE,
  from_rbatch = TRUE,
  model = 1
)
}
\arguments{
\item{eset}{The input expression matrix.}

\item{pdata}{additional patient data to include in the analysis (default is NULL).}

\item{id_pdata}{The identifier column name in pdata (default is "ID").}

\item{scale}{A logical value indicating whether to scale the data (default is FALSE).}

\item{check_eset}{A logical value indicating whether to remove outlier genes from the input expression matrix (default is FALSE).}

\item{from_rbatch}{A logical value indicating whether the sample was processed by the rbatch pipeline (default is TRUE).}

\item{model}{option = 1, or 2}
}
\description{
This function uses the LIRA model to predict new data and return the results.
}
