% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_model2.R
\name{lira_model2}
\alias{lira_model2}
\title{Title}
\usage{
lira_model2(
  eset,
  org = "hsa",
  pdata = NULL,
  id_pdata = "ID",
  scale = TRUE,
  plot = FALSE,
  ref = TRUE,
  loop = TRUE,
  check_eset = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{org}{'hsa' or 'mus'}

\item{ref}{}

\item{loop}{}

\item{return_all}{}
}
\description{
Title
}
