% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_rnaseq_pipline.R
\name{lira_rnaseq_pipline}
\alias{lira_rnaseq_pipline}
\title{Title}
\usage{
lira_rnaseq_pipline(
  pat_id,
  eset = NULL,
  data_type = "tpm",
  model = 1,
  palette = "nrc",
  path1 = NULL,
  path2 = NULL,
  data_path = "E:/03-NSCLC/19-NSCLC-LIRA/4-analysis/0-data/"
)
}
\arguments{
\item{data_path}{}
}
\description{
Title
}
\examples{
需要设置新数据的关键变量，否则后续画图会报错
寻找最佳的cutoff值
}
