% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lira_score_location.R
\name{lira_score_location}
\alias{lira_score_location}
\title{Location of lira_score of each patient}
\usage{
lira_score_location(
  score,
  pat_id,
  id_score = "ID",
  col_score = "riskscore",
  best_cutoff = NULL,
  ref_score = NULL,
  palette = "nrc",
  cols = NULL,
  palette_line = "jama",
  showplot = TRUE,
  path = NULL,
  panel = "OS",
  bins_width = 33,
  index = NULL,
  title = "LIRA score"
)
}
\arguments{
\item{score}{score}

\item{pat_id}{}

\item{id_score}{patient identifier}

\item{ref_score}{reference score of plot}

\item{palette}{palette of response}

\item{cols}{manual colors for objective response}

\item{palette_line}{palette of line}

\item{showplot}{default is FALSE}

\item{path}{path to save result}

\item{panel}{option = `OS`, `PFS`}
}
\description{
Location of lira_score of each patient
}
