% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rf_cox_model.R
\name{predict_rf_model_cox}
\alias{predict_rf_model_cox}
\title{Title}
\usage{
predict_rf_model_cox(
  sur_model,
  eset_new,
  pdata_new = NULL,
  feas = NULL,
  prefix = c("-", "\\\\:"),
  id_pdata_new = "ID",
  na_replace = 1
)
}
\arguments{
\item{na_replace}{}
}
\description{
Title
}
